Add files via upload
Digital Light Sensor- BH1750FVI. 

This sensor measures light intensity. It comes with a 16-bit Analog-Digital Converter and it has a I2C interface, which has a advantage that it fairly trivially implemented with the software.

The sensor implementation also has a advantage that it has been interfaced with ShunyaInterfaces. Shunya Interfaces makes the whole interfacing with the sensor very simple as all one would have to do is to use Shunya Interface library and the code is reduced by a significant part to just a single line.